+++ usecase/user/interface.go
@@ -6,16 +6,16 @@ import (
 
 //Reader interface
 type Reader interface {
-	Get(id int) (*entity.User, error)
+	Get(id int64) (*entity.User, error)
 	Search(query string) ([]*entity.User, error)
 	List() ([]*entity.User, error)
 }
 
 //Writer user writer
 type Writer interface {
-	Create(e *entity.User) (int, error)
+	Create(e *entity.User) (int64, error)
 	Update(e *entity.User) error
-	Delete(id int) error
+	Delete(id int64) error
 }
 
 //Repository interface
@@ -26,10 +26,10 @@ type Repository interface {
 
 //UseCase interface
 type UseCase interface {
-	GetUser(id int) (*entity.User, error)
+	GetUser(id int64) (*entity.User, error)
 	SearchUsers(query string) ([]*entity.User, error)
 	ListUsers() ([]*entity.User, error)
-	CreateUser(email, password, firstName, lastName string) (int, error)
+	CreateUser(email, password, firstName, lastName string) (int64, error)
 	UpdateUser(e *entity.User) error
-	DeleteUser(id int) error
+	DeleteUser(id int64) error
 }
